agentes-ia/dispatcher/requirements.txt
fastapi
uvicorn
requests
google-cloud-logging
